Forms: Extract RatingIcon to shared module
Move RatingIcon component from input-rating/edit.js to a dedicated
rating-icon.js file to allow importing without editor dependencies.
